1. 将 macOS 与 Windows 自动配置代理模块分开 2. 引入函数 AutoSetProxyOnMac 用法： AutoSetProxyOnMac(flag bool, portHttpWithPac int32, portSocks5 int32) 当 flag == true 是配置代理，portHttpWithPac 为带分流规则的 http 代理端口， portSocks5 为 socks5 代理端口
